+++ packages/components/molecules/f-mega-modal/CHANGELOG.md
@@ -4,6 +4,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+v0.6.1
+------------------------------
+*May 5, 2021*
+
+### Removed
+- `u-overlay` css style definition as should come from `fozzie`
+
+
 v0.6.0
 ------------------------------
 *March 5, 2021*