+++ packages/components/organisms/f-registration/CHANGELOG.md
@@ -3,6 +3,15 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+
+v0.51.1
+------------------------------
+*May 5, 2021*
+
+### Removed
+- `is-visuallyHidden` css style definition as should come from `fozzie`
+
+
 v0.51.0
 ------------------------------
 *April 15, 2021*